Refactor version extraction in CI workflow
Moved and modified the 'Extract version from .csproj and append -preview' step in `CI-development.yml`. The version extraction now occurs after the 'Restore dependencies' step and before the 'Build the solution (Release)' step. The extracted version is appended with '-preview', set as an environment variable using `$GITHUB_ENV`, and as an output variable using `::set-output`. The build step for packaging now uses this version with the '-preview' suffix for the `PackageVersion` property.